+++ IDisposableAnalyzers.Test/IDISP001DisposeCreatedTests/CodeFix.AddUsing.AddUsingDeclaration.cs
@@ -1,8 +1,10 @@
 ﻿namespace IDisposableAnalyzers.Test.IDISP001DisposeCreatedTests
 {
     using Gu.Roslyn.Asserts;
+
     using Microsoft.CodeAnalysis.CodeFixes;
     using Microsoft.CodeAnalysis.Diagnostics;
+
     using NUnit.Framework;
 
     public static partial class CodeFix
@@ -12,6 +14,19 @@
             private static readonly DiagnosticAnalyzer Analyzer = new LocalDeclarationAnalyzer();
             private static readonly ExpectedDiagnostic ExpectedDiagnostic = ExpectedDiagnostic.Create(Descriptors.IDISP001DisposeCreated);
             private static readonly CodeFixProvider Fix = new AddUsingFix();
+
+            private const string Disposable = @"
+namespace N
+{
+    using System;
+
+    public class Disposable : IDisposable
+    {
+        public void Dispose()
+        {
+        }
+    }
+}";
 
             [Test]
             public static void LocalToUsingDeclaration()
@@ -546,6 +561,42 @@ namespace N
 }";
                 RoslynAssert.CodeFix(Analyzer, Fix, ExpectedDiagnostic, before, after, fixTitle: "Add using to end of block.");
                 RoslynAssert.FixAll(Analyzer, Fix, ExpectedDiagnostic, before, after, fixTitle: "Add using to end of block.");
+            }
+
+            [TestCase("System.Activator.CreateInstance<Disposable>()")]
+            [TestCase("(Disposable)System.Activator.CreateInstance(typeof(Disposable))")]
+            [TestCase("(Disposable)constructorInfo.Invoke(null)")]
+            public static void Reflection(string expression)
+            {
+                var before = @"
+namespace N
+{
+    using System.Reflection;
+
+    public class C
+    {
+        public static void M(ConstructorInfo constructorInfo)
+        {
+            ↓var disposable = Activator.CreateInstance<Disposable>();
+        }
+    }
+}".AssertReplace("Activator.CreateInstance<Disposable>()", expression);
+
+                var after = @"
+namespace N
+{
+    using System.Reflection;
+
+    public class C
+    {
+        public static void M(ConstructorInfo constructorInfo)
+        {
+            using var disposable = Activator.CreateInstance<Disposable>();
+        }
+    }
+}".AssertReplace("Activator.CreateInstance<Disposable>()", expression);
+                RoslynAssert.CodeFix(Analyzer, Fix, ExpectedDiagnostic, new[] { Disposable, before }, after, fixTitle: "using");
+                RoslynAssert.FixAll(Analyzer, Fix, ExpectedDiagnostic, new[] { Disposable, before }, after, fixTitle: "using");
             }
         }
     }