+++ IDisposableAnalyzers/Helpers/KnownSymbols/KnownSymbol.cs
@@ -8,6 +8,7 @@ namespace IDisposableAnalyzers
         internal static readonly QualifiedType Void = Create("System.Void", "void");
         internal static readonly QualifiedType Object = Create("System.Object", "object");
         internal static readonly QualifiedType Boolean = Create("System.Boolean", "bool");
+        internal static readonly QualifiedType Activator = Create("System.Activator");
         internal static readonly QualifiedType Func = Create("System.Func");
         internal static readonly QualifiedType BinaryReader = Create("System.IO.BinaryReader");
         internal static readonly QualifiedType BinaryWriter = Create("System.IO.BinaryWriter");
@@ -19,6 +20,7 @@ namespace IDisposableAnalyzers
         internal static readonly QualifiedType DeflateStream = Create("System.IO.Compression.DeflateStream");
         internal static readonly QualifiedType GZipStream = Create("System.IO.Compression.GZipStream");
         internal static readonly QualifiedType StreamMemoryBlockProvider = Create("System.Reflection.Internal.StreamMemoryBlockProvider");
+        internal static readonly QualifiedType ConstructorInfo = Create("System.Reflection.ConstructorInfo");
         internal static readonly QualifiedType Attachment = Create("System.Net.Mail.Attachment");
 
         internal static readonly TupleType Tuple = new TupleType();