+++ ValidCode.NetCore/Reflection/Cases.cs
@@ -1,13 +1,16 @@
 ﻿namespace ValidCode.NetCore.Reflection
 {
     using System;
+    using System.Text;
+
     using Microsoft.Extensions.DependencyInjection;
 
     public class Cases
     {
         public static void ActivatorCreateInstanceOfDisposable(IServiceProvider provider)
         {
-            using var disposable = ActivatorUtilities.CreateInstance<Disposable>(provider);
+            using var disposable = Microsoft.Extensions.DependencyInjection.ActivatorUtilities.CreateInstance<Disposable>(provider);
+            var builder = Microsoft.Extensions.DependencyInjection.ActivatorUtilities.CreateInstance<StringBuilder>(provider);
         }
     }
 }